Revert "Rework recovery"
