+++ swap/actions.go
@@ -309,11 +309,7 @@ func (w *AwaitPaymentOrCsvAction) Execute(services *SwapServices, swap *SwapData
 		return swap.HandleError(err)
 	}
 
-	err = onchain.AddWaitForCsvTx(swap.GetId().String(), swap.OpeningTxBroadcasted.TxId, swap.OpeningTxBroadcasted.ScriptOut, swap.StartingBlockHeight, wantScript)
-	if err != nil {
-		// We only throw an error if the csv is already spent
-		return Event_AlreadyClaimed
-	}
+	onchain.AddWaitForCsvTx(swap.GetId().String(), swap.OpeningTxBroadcasted.TxId, swap.OpeningTxBroadcasted.ScriptOut, swap.StartingBlockHeight, wantScript)
 	return NoOp
 }
 
@@ -341,11 +337,7 @@ func (w *AwaitCsvAction) Execute(services *SwapServices, swap *SwapData) EventTy
 		return swap.HandleError(err)
 	}
 
-	err = onchain.AddWaitForCsvTx(swap.GetId().String(), swap.OpeningTxBroadcasted.TxId, swap.OpeningTxBroadcasted.ScriptOut, swap.StartingBlockHeight, wantScript)
-	if err != nil {
-		// We only throw an error if the csv is already spent
-		return Event_AlreadyClaimed
-	}
+	onchain.AddWaitForCsvTx(swap.GetId().String(), swap.OpeningTxBroadcasted.TxId, swap.OpeningTxBroadcasted.ScriptOut, swap.StartingBlockHeight, wantScript)
 	return NoOp
 }
 
@@ -659,14 +651,8 @@ func (t *AwaitTxConfirmationAction) Execute(services *SwapServices, swap *SwapDa
 	if err != nil {
 		return swap.HandleError(err)
 	}
-
+	txWatcher.AddWaitForConfirmationTx(swap.GetId().String(), swap.OpeningTxBroadcasted.TxId, swap.OpeningTxBroadcasted.ScriptOut, swap.StartingBlockHeight, wantScript)
 	log.Debugf("Await confirmation for tx with id: %s on swap %s", swap.OpeningTxBroadcasted.TxId, swap.GetId().String())
-	err = txWatcher.AddWaitForConfirmationTx(swap.GetId().String(), swap.OpeningTxBroadcasted.TxId, swap.OpeningTxBroadcasted.ScriptOut, swap.StartingBlockHeight, wantScript)
-	if err != nil {
-		// We only throw an error if the csv is already spent
-		return Event_AlreadyClaimed
-	}
-
 	return NoOp
 }
 
@@ -675,23 +661,9 @@ type ValidateTxAndPayClaimInvoiceAction struct{}
 
 func (p *ValidateTxAndPayClaimInvoiceAction) Execute(services *SwapServices, swap *SwapData) EventType {
 	lc := services.lightning
-	txWatcher, _, validator, err := services.getOnChainServices(swap.GetChain())
+	_, _, validator, err := services.getOnChainServices(swap.GetChain())
 	if err != nil {
 		return swap.HandleError(err)
-	}
-
-	// Check if we are outside of our csv safety limit. This can happen on
-	// restart. We do NOT want to continue if we fail here.
-	if swap.StartingBlockHeight > 0 {
-		now, err := txWatcher.GetBlockHeight()
-		if err != nil {
-			return swap.HandleError(err)
-		}
-
-		if now >= swap.StartingBlockHeight+(validator.GetCSVHeight()/2) {
-			err := fmt.Errorf("exceeded csv limit")
-			return swap.HandleError(err)
-		}
 	}
 
 	// todo get opening tx hex
