+++ swap/services.go
@@ -51,8 +51,8 @@ type LightningClient interface {
 }
 
 type TxWatcher interface {
-	AddWaitForConfirmationTx(swapId, txId string, vout, startingHeight uint32, scriptpubkey []byte) error
-	AddWaitForCsvTx(swapId, txId string, vout uint32, startingHeight uint32, scriptpubkey []byte) error
+	AddWaitForConfirmationTx(swapId, txId string, vout, startingHeight uint32, scriptpubkey []byte)
+	AddWaitForCsvTx(swapId, txId string, vout uint32, startingHeight uint32, scriptpubkey []byte)
 	AddConfirmationCallback(func(swapId string, txHex string) error)
 	AddCsvCallback(func(swapId string) error)
 	GetBlockHeight() (uint32, error)
